puppeteer sandbox
This is extremely frustrating. IMO Github / Google
should not have made this change without providing
a solution. Puppeteer's message is useless for github
actions. It assumes you control the server. The majority
of projects do not.

So, sandbox off. If that's bad for github too fucking bad.
If Google/Github want us to do the right thing then they
should either make the right thing happen by default. Or,
provide docs on how to do it. The docs provided are insufficient.